restructure and remove previous verions
