+++ pkg/schema/introspection.go
@@ -1,0 +1,7 @@
+package schema
+
+import (
+	"context"
+)
+
+type IntrospectionRule func(context.Context) bool
